test pendingTx case
